move create_coop_view-provider in correct module
